+++ CHANGES.rst
@@ -49,6 +49,7 @@ mio 0.0.6.dev
     c()
     
 - ``Object evalArg`` should evaluate the argument with context as the receiver.
+- Added ``List __getitem__`` and ``List __len__`` methods.
 
 
 mio 0.0.5 (2013-10-29)